+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v 0.1.0 (?)
 Changes in this release:
-- Added logging of deployment failure when awaited service instance reaches bad state (#49).
-
+- Added update method to the remote orchestrator.
+- Added logging for deployment failure (#35) and more explanations on failures overall.
 - Add support to override the environment settings that are set after a clean
 - Expose the noclean boolean in the object returned by remote_orchestator fixtures for other fixtures to hook into
 - Fix issue #42 where the fixture fails if a compile is in progress